+++ src/ResourceManager/Compute/Commands.Compute/Strategies/ComputeRp/AvailabilitySetStrategy.cs
@@ -12,43 +12,38 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.Commands.Common.Strategies;
+using Microsoft.Azure.Commands.Common.Strategies.Compute;
 using Microsoft.Azure.Commands.Compute.Strategies.ResourceManager;
 using Microsoft.Azure.Management.Compute;
 using Microsoft.Azure.Management.Compute.Models;
 using Microsoft.Azure.Management.Internal.Resources.Models;
 
-namespace Microsoft.Azure.Commands.Common.Strategies.Compute
+namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
 {
-    static class ManagedDiskStrategy
+    static class AvailabilitySetStrategy
     {
-        public static ResourceStrategy<Disk> Strategy { get; }
+        public static ResourceStrategy<AvailabilitySet> Strategy { get; }
             = ComputePolicy.Create(
-                provider: "disks",
-                getOperations: client => client.Disks,
+                type: "availability set",
+                provider: "availabilitySets",
+                getOperations: client => client.AvailabilitySets,
                 getAsync: (o, p) => o.GetAsync(
                     p.ResourceGroupName, p.Name, p.CancellationToken),
                 createOrUpdateAsync: (o, p) => o.CreateOrUpdateAsync(
                     p.ResourceGroupName, p.Name, p.Model, p.CancellationToken),
-                createTime: d => 120);
+                createTime: c => 1);
 
-        public static ResourceConfig<Disk> CreateManagedDiskConfig(
-            this ResourceConfig<ResourceGroup> resourceGroup,
-            string name,
-            string sourceUri)
+        public static ResourceConfig<AvailabilitySet> CreateAvailabilitySetConfig(
+            this ResourceConfig<ResourceGroup> resourceGroup, string name)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
-                createModel: subscription => new Disk
+                createModel: subscription => new AvailabilitySet
                 {
-                    Sku = new DiskSku
-                    {
-                        Name = StorageAccountTypes.PremiumLRS
-                    },
-                    CreationData = new CreationData
-                    {
-                        CreateOption = DiskCreateOption.Import,
-                        SourceUri = sourceUri
-                    }
+                    Sku = new Azure.Management.Compute.Models.Sku {  Name = "Aligned" },
+                    PlatformFaultDomainCount = 2,
+                    PlatformUpdateDomainCount = 2,
                 });
     }
 }