+++ src/ResourceManager/Compute/Commands.Compute/Strategies/ComputeRp/VirtualMachineScaleSetStrategy.cs
@@ -20,8 +20,9 @@ using Microsoft.Azure.Management.Internal.Network.Version2017_10_01.Models;
 using System.Linq;
 using System.Collections.Generic;
 using System;
+using Microsoft.Azure.Commands.Common.Strategies;
 
-namespace Microsoft.Azure.Commands.Common.Strategies.Compute
+namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
 {
     public static class VirtualMachineScaleSetStrategy
     {
@@ -41,10 +42,9 @@ namespace Microsoft.Azure.Commands.Common.Strategies.Compute
             NestedResourceConfig<Subnet, VirtualNetwork> subnet,
             IEnumerable<NestedResourceConfig<FrontendIPConfiguration, LoadBalancer>> frontendIpConfigurations,
             NestedResourceConfig<BackendAddressPool, LoadBalancer> backendAdressPool,
-            bool isWindows,
+            Func<ImageAndOsType> getImageAndOsType,
             string adminUsername,
             string adminPassword,
-            Image image,
             string vmSize,
             int instanceCount,
             UpgradeMode? upgradeMode)
@@ -53,6 +53,7 @@ namespace Microsoft.Azure.Commands.Common.Strategies.Compute
                 name: name,
                 createModel: subscriptionId =>
                 {
+                    var imageAndOsType = getImageAndOsType();
                     var vmss = new VirtualMachineScaleSet()
                     {
                         Zones = frontendIpConfigurations
@@ -62,17 +63,12 @@ namespace Microsoft.Azure.Commands.Common.Strategies.Compute
                             .Where(z => z != null)
                             .ToList(),
 
-                        UpgradePolicy = upgradeMode.HasValue ?
-                            new UpgradePolicy
-                            {
-                                Mode = upgradeMode
-                            }
-                            : new UpgradePolicy
-                            {
-                                Mode = UpgradeMode.Manual
-                            },
+                        UpgradePolicy =new UpgradePolicy
+                        {
+                            Mode = upgradeMode ?? UpgradeMode.Manual
+                        },
 
-                        Sku = new Microsoft.Azure.Management.Compute.Models.Sku()
+                        Sku = new Azure.Management.Compute.Models.Sku()
                         {
                             Capacity = instanceCount,
                             Name = vmSize,
@@ -83,21 +79,19 @@ namespace Microsoft.Azure.Commands.Common.Strategies.Compute
                     vmss.VirtualMachineProfile.OsProfile = new VirtualMachineScaleSetOSProfile
                     {
                         ComputerNamePrefix = name.Substring(0, Math.Min(name.Length, 9)),
-                        WindowsConfiguration = isWindows ? new WindowsConfiguration { } : null,
-                        LinuxConfiguration = isWindows ? null : new LinuxConfiguration(),
+                        WindowsConfiguration = imageAndOsType.OsType == OperatingSystemTypes.Windows 
+                            ? new WindowsConfiguration()
+                            : null,
+                        LinuxConfiguration = imageAndOsType.OsType == OperatingSystemTypes.Linux 
+                            ? new LinuxConfiguration()
+                            : null,
                         AdminUsername = adminUsername,
                         AdminPassword = adminPassword,
                     };
 
                     vmss.VirtualMachineProfile.StorageProfile = new VirtualMachineScaleSetStorageProfile
                     {
-                        ImageReference = new ImageReference
-                        {
-                            Publisher = image.publisher,
-                            Offer = image.offer,
-                            Sku = image.sku,
-                            Version = image.version
-                        }
+                        ImageReference = imageAndOsType.Image
                     };
 
                     var ipConfig = new VirtualMachineScaleSetIPConfiguration
@@ -129,6 +123,6 @@ namespace Microsoft.Azure.Commands.Common.Strategies.Compute
                     return vmss;
                 },
                 dependencies: new IEntityConfig[] { subnet, backendAdressPool }
-                                                  .Concat(frontendIpConfigurations));
+                    .Concat(frontendIpConfigurations));
     }
 }