+++ src/ResourceManager/Compute/Commands.Compute/VirtualMachine/Operation/NewAzureVMCommand.cs
@@ -16,15 +16,13 @@ using AutoMapper;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.Common.Strategies;
-using Microsoft.Azure.Commands.Common.Strategies.Compute;
-using Microsoft.Azure.Commands.Common.Strategies.Network;
-using Microsoft.Azure.Commands.Common.Strategies.ResourceManager;
-using Microsoft.Azure.Commands.Common.Strategies.Templates;
 using Microsoft.Azure.Commands.Compute.Common;
 using Microsoft.Azure.Commands.Compute.Models;
 using Microsoft.Azure.Commands.Compute.StorageServices;
 using Microsoft.Azure.Commands.Compute.Strategies;
-using Microsoft.Azure.Commands.Compute.Strategies.Compute;
+using Microsoft.Azure.Commands.Compute.Strategies.ComputeRp;
+using Microsoft.Azure.Commands.Compute.Strategies.Network;
+using Microsoft.Azure.Commands.Compute.Strategies.ResourceManager;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Management.Compute;
 using Microsoft.Azure.Management.Compute.Models;
@@ -50,7 +48,6 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using CM = Microsoft.Azure.Management.Compute.Models;
-using SM = Microsoft.Azure.Management.Storage.Models;
 
 namespace Microsoft.Azure.Commands.Compute
 {
@@ -244,48 +241,7 @@ namespace Microsoft.Azure.Commands.Compute
             PublicIpAddressName = PublicIpAddressName ?? Name;
             SecurityGroupName = SecurityGroupName ?? Name;
 
-            bool isWindows;
-            Commands.Common.Strategies.Compute.Image image = null;
-            if (ImageName.Contains(':'))
-            {
-                var imageArray = ImageName.Split(':');
-                if (imageArray.Length != 4)
-                {
-                    throw new Exception("Invalid ImageName");
-                }
-                image = new Commands.Common.Strategies.Compute.Image
-                {
-                    publisher = imageArray[0],
-                    offer = imageArray[1],
-                    sku = imageArray[2],
-                    version = imageArray[3],
-                };
-                isWindows = image.publisher.ToLower() == "MicrosoftWindowsServer".ToLower();
-            }
-            else if (!string.IsNullOrEmpty(DiskFile))
-            {
-                // disk file parameter set requires the OS type input
-                isWindows = !Linux;
-            }
-            else
-            {
-                // get image
-                var osTypeAndImage = Images
-                    .Instance
-                    .SelectMany(osAndMap => osAndMap
-                        .Value
-                        .Where(nameAndImage => nameAndImage.Key.ToLower() == ImageName.ToLower())
-                        .Select(nameAndImage => new
-                        {
-                            OsType = osAndMap.Key,
-                            Image = nameAndImage.Value
-                        }))
-                    .FirstOrDefault();
-                image = osTypeAndImage.Image;
-                isWindows = osTypeAndImage.OsType == "Windows";
-            }
-            
-            OpenPorts = OpenPorts ?? (isWindows ? new[] { 3389, 5985 } : new[] { 22 });
+            var imageAndOsType = new ImageAndOsType(OperatingSystemTypes.Windows, null);
 
             var passwordValue = new NetworkCredential(string.Empty, Credential.Password).Password;
 
@@ -301,24 +257,25 @@ namespace Microsoft.Azure.Commands.Compute
                 allocationMethod: AllocationMethod);
             var networkSecurityGroup = resourceGroup.CreateNetworkSecurityGroupConfig(
                 name: SecurityGroupName,
-                openPorts: OpenPorts);
+                openPorts: OpenPorts,
+                getOsType: () => imageAndOsType.OsType);
             var networkInterface = resourceGroup.CreateNetworkInterfaceConfig(
                 Name, subnet, publicIpAddress, networkSecurityGroup);
 
             var availabilitySet = AvailabilitySetName == null 
                 ? null
-                : resourceGroup.CreateAvailabilitySetConfig(name: Name);
+                : resourceGroup.CreateAvailabilitySetConfig(name: AvailabilitySetName);
 
             ResourceConfig<VirtualMachine> virtualMachine = null;
-            if (image != null)
+            if (DiskFile == null)
             {
                 virtualMachine = resourceGroup.CreateVirtualMachineConfig(
                     name: Name,
                     networkInterface: networkInterface,
-                    isWindows: isWindows,
+                    getImageAndOsType: () => imageAndOsType,
                     adminUsername: Credential.UserName,
                     adminPassword: password,
-                    image: image,
+
                     size: Size,
                     availabilitySet: availabilitySet);
             }
@@ -335,15 +292,21 @@ namespace Microsoft.Azure.Commands.Compute
                         Name = ResourceGroupName
                     });
                 }
+                imageAndOsType = new ImageAndOsType(
+                    Linux ? OperatingSystemTypes.Linux : OperatingSystemTypes.Windows,
+                    null);
                 var storageClient =
                         AzureSession.Instance.ClientFactory.CreateArmClient<StorageManagementClient>(DefaultProfile.DefaultContext,
                             AzureEnvironment.Endpoint.ResourceManager);
-                var st1 = storageClient.StorageAccounts.Create(ResourceGroupName, Name, new StorageAccountCreateParameters
+                var st1 = storageClient.StorageAccounts.Create(
+                    ResourceGroupName,
+                    Name,
+                    new StorageAccountCreateParameters
                 {
 #if !NETSTANDARD
                     AccountType = AccountType.PremiumLRS,
 #else
-                    Sku = new SM.Sku
+                    Sku = new Microsoft.Azure.Management.Storage.Models.Sku
                     {
                         Name = SkuName.PremiumLRS
                     },
@@ -399,7 +362,7 @@ namespace Microsoft.Azure.Commands.Compute
                 virtualMachine = resourceGroup.CreateVirtualMachineConfig(
                     name: Name,
                     networkInterface: networkInterface,
-                    isWindows: isWindows,
+                    osType: imageAndOsType.OsType,
                     disk: disk,
                     size: Size,
                     availabilitySet: availabilitySet);
@@ -410,14 +373,7 @@ namespace Microsoft.Azure.Commands.Compute
             // get current Azure state
             var current = await virtualMachine.GetStateAsync(client, new CancellationToken());
 
-            if (Location == null)
-            {
-                Location = current.GetLocation(virtualMachine);
-                if (Location == null)
-                {
-                    Location = "eastus";
-                }
-            }
+            Location = current.UpdateLocation(Location, virtualMachine);
 
             // generate a domain name label if it's not specified.
             DomainNameLabel = await PublicIPAddressStrategy.UpdateDomainNameLabelAsync(
@@ -427,6 +383,11 @@ namespace Microsoft.Azure.Commands.Compute
                 client: client);
 
             var fqdn = PublicIPAddressStrategy.Fqdn(DomainNameLabel, Location);
+
+            if (DiskFile == null)
+            {
+                imageAndOsType = await client.UpdateImageAndOsTypeAsync(ImageName, Location);
+            }
 
             // template
             if (AsArmTemplate)
@@ -523,20 +484,19 @@ namespace Microsoft.Azure.Commands.Compute
                         new ShouldProcess(asyncCmdlet),
                         asyncCmdlet.ReportTaskProgress);
 
-                var result = newState.Get(virtualMachine);
-                if (result == null)
-                {
-                    result = current.Get(virtualMachine);
-                }
-                if (result != null)
-                {
-                    var psResult = ComputeAutoMapperProfile.Mapper.Map<PSVirtualMachine>(result);
-                    psResult.FullyQualifiedDomainName = fqdn;
-                    asyncCmdlet.WriteVerbose(isWindows
-                        ? "Use 'mstsc /v:" + fqdn + "' to connect to the VM."
-                        : "Use 'ssh " + Credential.UserName + "@" + fqdn + "' to connect to the VM.");
-                    asyncCmdlet.WriteObject(psResult);
-                }
+            var result = newState.Get(virtualMachine);
+            if (result == null)
+            {
+                result = current.Get(virtualMachine);
+            }
+            if (result != null)
+            {
+                var psResult = ComputeAutoMapperProfile.Mapper.Map<PSVirtualMachine>(result);
+                psResult.FullyQualifiedDomainName = fqdn;
+                asyncCmdlet.WriteVerbose(imageAndOsType.OsType == OperatingSystemTypes.Windows
+                    ? "Use 'mstsc /v:" + fqdn + "' to connect to the VM."
+                    : "Use 'ssh " + Credential.UserName + "@" + fqdn + "' to connect to the VM.");
+                asyncCmdlet.WriteObject(psResult);
             }
         }