+++ src/ResourceManager/Compute/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added `AvailabilitySetName` parameter to the simplified parameterset of `New-AzureRmVm`.
 * Corrected usage of `Login-AzureRmAccount` to use `Connect-AzureRmAccount`
 * User assigned identity support for VM and VM scale set
 - IdentityType and IdentityId parameters are added to New-AzureRmVMConfig, New-AzureRmVmssConfig, Update-AzureRmVM and Update-AzureRmVmss
@@ -31,6 +32,7 @@
 * Obsoleted -Tags in favor of -Tag for New-AzureRmVm and Update-AzureRmVm
 * Fixed Get-AzureRmComputeResourceSku cmdlet when Zone is included in restriction.
 * Updated Diagnostics Agent configuration schema for Azure Monitor sink support.
+* Fixed a non-existing resource group creation problem for New-AzureRmVM
 
 ## Version 4.1.1
 * Get-AzureRmComputeResourceSku shows zone information.
@@ -75,7 +77,7 @@
     - New cmdelt: 'Get-AzureRmVmssVMDiskEncryptionStatus' shows the disk encryption status of VMs in a VM scale set
 
 ## Version 3.3.1
-* 
+*
 ## Version 3.3.0
 * Set-AzureRmVMAEMExtension: Add support for new Premium Disk sizes
 * Set-AzureRmVMAEMExtension: Add support for M series
@@ -92,7 +94,7 @@
   - New-AzureRmSnapshot
   - Update-AzureRmDisk
   - Update-AzureRmSnapshot
-  
+
 ## Version 3.2.0
 * Storage account type support for Image disk:
     - 'StorageAccountType' parameter is added to Set-AzureRmImageOsDisk and Add-AzureRmImageDataDisk